requirements.txt
google-api-python-client==2.115.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
spacy==3.7.4
transformers==4.36.2
torch==2.1.2
flask==3.0.0
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2